Fix for local player info in build 1048
Fixes bug from 1048 where the local player's bone info would be
removed on scene change. For some reason OnPlayerLeft is getting
called for the local player right after joining a world now. Added a
simple check to disable that from removing the local player. Since the
mod resets on world change, it shouldn't be needed for the local
player anyway.